test: add tests for StartingMoneyActivity and StartingMoneyScreen
